# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NFL Survivor Pool API",
    "version": "1.0.0",
    "description": "API for NFL Survivor Pool Web App"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Request": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "The auto-generated id of the request"
          },
          "user": {
            "type": "string",
            "description": "The id of the user who made the request"
          },
          "pool": {
            "type": "string",
            "description": "The id of the pool the request is for"
          },
          "numberOfEntries": {
            "type": "integer",
            "description": "The number of entries requested"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "payment_pending"
            ],
            "description": "The current status of the request"
          },
          "totalAmount": {
            "type": "number",
            "description": "The total amount to be paid for the entries"
          },
          "paymentStatus": {
            "type": "string",
            "enum": [
              "pending",
              "confirmed",
              "failed"
            ],
            "description": "The status of the payment for this request"
          },
          "transactionId": {
            "type": "string",
            "description": "The transaction id of the payment"
          },
          "paymentType": {
            "type": "string",
            "description": "The type of payment used"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "The date and time when the request was created"
          }
        }
      },
      "Entry": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "user": {
            "type": "string"
          },
          "pool": {
            "type": "string"
          },
          "request": {
            "type": "string"
          },
          "entryNumber": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "eliminated"
            ]
          },
          "eliminatedWeek": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/requests": {
      "post": {
        "summary": "Create a new request",
        "tags": [
          "Requests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "poolId",
                  "numberOfEntries"
                ],
                "properties": {
                  "poolId": {
                    "type": "string"
                  },
                  "numberOfEntries": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Request created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Request"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "Get all requests for the current user",
        "tags": [
          "Requests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of user requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Request"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/requests/{id}": {
      "get": {
        "summary": "Get a specific request",
        "tags": [
          "Requests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Request"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Request not found"
          }
        }
      }
    },
    "/api/v1/requests/{id}/confirm-payment": {
      "put": {
        "summary": "Confirm payment for a request",
        "tags": [
          "Requests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "transactionId",
                  "paymentType"
                ],
                "properties": {
                  "transactionId": {
                    "type": "string"
                  },
                  "paymentType": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment confirmed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Request"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Request not found"
          }
        }
      }
    },
    "/api/v1/requests/pool/{poolId}": {
      "get": {
        "summary": "Get all requests for a specific pool (Admin only)",
        "tags": [
          "Requests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "poolId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of pool requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Request"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/requests/{id}/approve": {
      "put": {
        "summary": "Approve a request (Admin only)",
        "tags": [
          "Requests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request approved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "request": {
                          "$ref": "#/components/schemas/Request"
                        },
                        "entries": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Entry"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Request not found"
          }
        }
      }
    },
    "/api/v1/requests/{id}/reject": {
      "put": {
        "summary": "Reject a request (Admin only)",
        "tags": [
          "Requests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request rejected successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Request"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Request not found"
          }
        }
      }
    }
  },
...
}